my_agent_project/agent_development/agent_requirements.txt
unitycatalog-langchain[databricks]==0.2.0
databricks-vectorsearch==0.56
databricks-langchain==0.5.1
langgraph==0.3.4 # 5.6.0
mlflow==3.1.1
databricks-agents==1.1.0